1.0 commit
It all functions.
There's some odd stuff with reset objects having some spasms on restart.
I could potentially add more techniques to the utensils themselves during sandwich cutting, but I'm not sure if that makes sense. I'll have to think about it, it shouldn't be too hard to add if so.