1.10
Made the hand highlight material set to 0 on start. Also added a check for the renderer to still exist, just in case.
Updated the material for the utensils to be just a bit lighter
Moved the sandwich ingredients to the main plate, rather than having them reach around more. It makes sense so that people don't have to reach for the utensils.